pyproject.toml
[project]
name = "spond-mcp-server"
version = "0.1.0"
description = "MCP server exposing the Spond API as tools for Windsurf/Cascade"
requires-python = ">=3.11"
dependencies = [
    "spond>=1.0.0",
    "mcp[cli]>=1.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
spond-mcp-server = "spond_mcp_server.server:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "integration: live API tests requiring real Spond credentials",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
